Don’t call gem
In `lib/compat/multi_json.rb` there is a call to `Kernel#gem`. Normally these things will be handled by Bundler so it is unclear why this is here.

This causes a problem when using https://github.com/iconara/puck since it doesn’t actually include gem specs but sets up required gems statically in a jar-file, causing the call to fail.

The solution is to simply remove that call.